requirements.txt
asgiref==3.11.0
autopep8==2.3.2
black==25.12.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distlib==0.4.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-extensions==4.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
filelock==3.20.3
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.16
idna==3.11
isort==7.0.0
librt==0.7.8
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.4.1
packaging==25.0
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
pyOpenSSL==25.3.0
python-decouple==3.8
pytokens==0.3.0
PyYAML==6.0.3
PyJWT==2.10.1
requests==2.32.5
strawberry-graphql==0.235.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
webencodings==0.5.1
Werkzeug==3.1.5
